pp_desktop_assistant/requirements.txt
openai
pyttsx3
SpeechRecognition
requests
pyautogui
pygame
plyer
beautifulsoup4
speedtest-cli
matplotlib
googletrans
gtts
playsound
pynput
pywhatkit
wikipedia
wolframalpha
pillow
tk
